+++ package.json
@@ -14,6 +14,7 @@
   },
   "dependencies": {
     "@hot-loader/react-dom": "^16.11.0",
+    "@types/zepto": "^1.0.30",
     "antd": "^4.0.4",
     "axios": "^0.19.2",
     "chroma-js": "^2.1.0",
